+++ src/mito2/src/worker.rs
@@ -54,10 +54,7 @@ use crate::error::{CreateDirSnafu, JoinSnafu, Result, WorkerStoppedSnafu};
 use crate::flush::{FlushScheduler, WriteBufferManagerImpl, WriteBufferManagerRef};
 use crate::memtable::MemtableBuilderProvider;
 use crate::metrics::{REGION_COUNT, WRITE_STALL_TOTAL};
-use crate::region::opener::RegionOpener;
-use crate::region::{
-    MitoRegion, MitoRegionRef, OpeningRegions, OpeningRegionsRef, RegionMap, RegionMapRef,
-};
+use crate::region::{MitoRegionRef, OpeningRegions, OpeningRegionsRef, RegionMap, RegionMapRef};
 use crate::request::{
     BackgroundNotify, DdlRequest, SenderDdlRequest, SenderWriteRequest, WorkerRequest,
 };
@@ -939,40 +936,6 @@ impl<S: LogStore> RegionWorkerLoop<S> {
         } else {
             let _ = sender.send(SetRegionRoleStateResponse::NotFound);
         }
-    }
-
-    /// Reopens a region.
-    pub(crate) async fn reopen_region(
-        &mut self,
-        region: &Arc<MitoRegion>,
-        is_mutable_empty: bool,
-        is_immutable_empty: bool,
-    ) -> Result<Arc<MitoRegion>> {
-        let region_id = region.region_id;
-        let manifest_version = region.manifest_ctx.manifest_version().await;
-        let flushed_entry_id = region.version_control.current().last_entry_id;
-        info!("Reopening the region: {region_id}, empty mutable: {is_mutable_empty}, empty immutable: {is_immutable_empty}, manifest version: {manifest_version}, flushed entry id: {flushed_entry_id}");
-        let reopened_region = Arc::new(
-            RegionOpener::new(
-                region_id,
-                region.region_dir(),
-                self.memtable_builder_provider.clone(),
-                self.object_store_manager.clone(),
-                self.purge_scheduler.clone(),
-                self.puffin_manager_factory.clone(),
-                self.intermediate_manager.clone(),
-                self.time_provider.clone(),
-            )
-            .cache(Some(self.cache_manager.clone()))
-            .options(region.version().options.clone())?
-            .skip_wal_replay(true)
-            .open(&self.config, &self.wal)
-            .await?,
-        );
-        debug_assert!(!reopened_region.is_writable());
-        self.regions.insert_region(reopened_region.clone());
-
-        Ok(reopened_region)
     }
 }
