+++ src/mito2/src/worker/handle_manifest.rs
@@ -148,7 +148,7 @@ impl<S: LogStore> RegionWorkerLoop<S> {
             }
         };
         let version = region.version();
-        if !version.memtables.mutable.is_empty() || !version.memtables.immutables().is_empty() {
+        if !version.memtables.is_empty() {
             warn!(
                 "Region {} memtables is not empty, which should not happen, manifest version: {}",
                 region.region_id, manifest.manifest_version
